Extract shared plain address handler construction
AddressHandlerFactory and AddressInputModule each built the same per-chain
handler list in hand-synced when blocks. Both now use plainAddressHandlers().
The AddressHandlerPure branch in AddressInputModule was unreachable (duplicate
when conditions - first match wins), so dropping it changes no behavior.
