+++ walletkit/src/main/java/io/horizontalsystems/walletkit/modules/address/AddressInputModule.kt
@@ -2,15 +2,9 @@ package io.horizontalsystems.walletkit.modules.address
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import io.horizontalsystems.bitcoincash.MainNetBitcoinCash
-import io.horizontalsystems.bitcoinkit.MainNet
 import io.horizontalsystems.walletkit.core.App
 import io.horizontalsystems.walletkit.core.utils.AddressUriParser
 import io.horizontalsystems.walletkit.entities.Address
-import io.horizontalsystems.dashkit.MainNetDash
-import io.horizontalsystems.ecash.MainNetECash
-import io.horizontalsystems.litecoinkit.MainNetLitecoin
-import io.horizontalsystems.marketkit.models.BlockchainType
 import io.horizontalsystems.marketkit.models.TokenQuery
 
 object AddressInputModule {
@@ -23,81 +17,8 @@ object AddressInputModule {
             val udnHandler = AddressHandlerUdn(tokenQuery, coinCode, App.appConfigProvider.udnApiKey)
             val addressParserChain = AddressParserChain(domainHandlers = listOf(ensHandler, udnHandler))
 
-            when (blockchainType) {
-                BlockchainType.Bitcoin -> {
-                    val network = MainNet()
-                    addressParserChain.addHandler(AddressHandlerBase58(network, blockchainType))
-                    addressParserChain.addHandler(AddressHandlerBech32(network, blockchainType))
-                }
-
-                BlockchainType.BitcoinCash -> {
-                    val network = MainNetBitcoinCash()
-                    addressParserChain.addHandler(AddressHandlerBase58(network, blockchainType))
-                    addressParserChain.addHandler(AddressHandlerBitcoinCash(network, blockchainType))
-                }
-
-                BlockchainType.ECash -> {
-                    val network = MainNetECash()
-                    addressParserChain.addHandler(AddressHandlerBase58(network, blockchainType))
-                    addressParserChain.addHandler(AddressHandlerBitcoinCash(network, blockchainType))
-                }
-
-                BlockchainType.Litecoin -> {
-                    val network = MainNetLitecoin()
-                    addressParserChain.addHandler(AddressHandlerBase58(network, blockchainType))
-                    addressParserChain.addHandler(AddressHandlerBech32(network, blockchainType))
-                }
-
-                BlockchainType.Dash -> {
-                    val network = MainNetDash()
-                    addressParserChain.addHandler(AddressHandlerBase58(network, blockchainType))
-                }
-
-                BlockchainType.Zcash -> {
-                    addressParserChain.addHandler(AddressHandlerZcash())
-                }
-                BlockchainType.Bitcoin,
-                BlockchainType.BitcoinCash,
-                BlockchainType.ECash,
-                BlockchainType.Litecoin,
-                BlockchainType.Dash,
-                BlockchainType.Zcash -> {
-                    addressParserChain.addHandler(AddressHandlerPure(blockchainType))
-                }
-                BlockchainType.Ethereum,
-                BlockchainType.BinanceSmartChain,
-                BlockchainType.Polygon,
-                BlockchainType.Avalanche,
-                BlockchainType.Optimism,
-                BlockchainType.Base,
-                BlockchainType.ZkSync,
-                BlockchainType.Gnosis,
-                BlockchainType.Fantom,
-                BlockchainType.ArbitrumOne -> {
-                    addressParserChain.addHandler(AddressHandlerEvm(blockchainType))
-                }
-                BlockchainType.Solana -> {
-                    addressParserChain.addHandler(AddressHandlerSolana())
-                }
-                BlockchainType.Tron -> {
-                    addressParserChain.addHandler(AddressHandlerTron())
-                }
-                BlockchainType.Ton -> {
-                    addressParserChain.addHandler(AddressHandlerTon())
-                }
-                BlockchainType.Stellar -> {
-                    addressParserChain.addHandler(AddressHandlerStellar())
-                }
-                BlockchainType.Thorchain -> {
-                    addressParserChain.addHandler(AddressHandlerThorchain())
-                }
-                BlockchainType.Monero -> {
-                    addressParserChain.addHandler(AddressHandlerMonero())
-                }
-                BlockchainType.Zano -> {
-                    addressParserChain.addHandler(AddressHandlerZano())
-                }
-                is BlockchainType.Unsupported -> Unit
+            plainAddressHandlers(blockchainType).forEach {
+                addressParserChain.addHandler(it)
             }
 
             val addressUriParser = AddressUriParser(blockchainType, tokenQuery.tokenType)
